requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pandas==2.2.3
openpyxl==3.1.5
firebase-admin==6.7.0
gunicorn==23.0.0
reportlab==4.4.0
xlsxwriter==3.2.2
